requirements.txt
# Core
streamlit>=1.35.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
uuid>=1.30
